Allow Tyger to run in Docker (#103)

+++ scripts/az-login-from-host.sh
@@ -8,8 +8,12 @@
 
 set -euo pipefail
 
-devcontainer_id=$(cat /etc/hostname)
-devcontainer_image=$(docker inspect "$devcontainer_id" | jq -r '.[0].Image')
+if [[ -n "${WSL_DISTRO_NAME:-}" ]]; then
+    echo "Reusing Azure CLI context cache is not supported from WSL." >&2
+    exit 1
+fi
+
+devcontainer_image=$(docker ps --filter label=devcontainer.metadata --format json | jq -r '.Image' | head -n 1)
 container_id=$(docker run -d -u "$USER" --mount "source=${DEVCONTAINER_HOST_HOME}/.azure/,target=/home/${USER}/.azure,type=bind,readonly" "$devcontainer_image" 2>/dev/null || true)
 if [[ -n "$container_id" ]]; then
     trap 'docker rm -f "$container_id" 2&> /dev/null' EXIT
